+++ demisto_sdk/commands/common/tests/conf_test.py
@@ -1,4 +1,5 @@
 import pytest
+from demisto_sdk.commands.common.constants import FileType
 from demisto_sdk.commands.common.hook_validations.conf_json import \
     ConfJsonValidator
 
@@ -87,3 +88,94 @@ def test_is_valid_conf_json_negative_sanity_check():
 
     assert validator.is_valid_conf_json() is False, \
         "The conf validator didn't find the missing description sections although they don't exist"
+
+
+INTEGRATION_IS_SKIPPED_TEST_INPUTS = [
+    ({"tests": [{"integrations": "SomeIntegration", "playbookID": "SomeTestPlaybook"}],
+      "skipped_tests": {"SomeTestPlaybook": "Some Issue"}},
+     False),
+    ({"tests": [{"integrations": "SomeIntegration", "playbookID": "SomeTestPlaybook"}],
+      "skipped_tests": {"SomeOtherTestPlaybook": "Some Issue"}},
+     True),
+    ({"tests": [{"integrations": "SomeIntegration", "playbookID": "SomeTestPlaybook"},
+                {"integrations": ["SomeIntegration", "SomeOtherIntegration"],
+                 "playbookID": "SomeCombinedTestPlaybook"}],
+      "skipped_tests": {"SomeTestPlaybook": "Some Issue"}},
+     True)
+]
+
+
+@pytest.mark.parametrize("conf_dict, answer", INTEGRATION_IS_SKIPPED_TEST_INPUTS)
+def test_integration_has_unskipped_test_playbook(mocker, conf_dict, answer):
+    """
+    Given:
+        - An integration.
+        - conf file with configurations for the integration.
+
+    When: running is_valid_file_in_conf_json specifically on integration.
+
+    Then: Validate the integration has at least one unskipped test playbook.
+    """
+    mocker.patch.object(ConfJsonValidator, 'load_conf_file', return_value=conf_dict)
+
+    validator = ConfJsonValidator()
+    current = {"commonfields": {"id": "SomeIntegration"}, "tests": ["SomeTestPlaybook"]}
+
+    assert validator.is_valid_file_in_conf_json(current_file=current,
+                                                file_type=FileType.INTEGRATION,
+                                                file_path="SomeFilePath") is answer
+
+
+SCRIPT_IS_SKIPPED_TEST_INPUTS = [
+    (["SomeTestPlaybook"],
+     {"skipped_tests": {"SomeTestPlaybook": "Some Issue"}},
+     False),
+    (["SomeTestPlaybook"],
+     {"skipped_tests": {"SomeOtherTestPlaybook": "Some Issue"}},
+     True),
+    (["SomeTestPlaybook", "SomeSecondTestPlaybook"],
+     {"skipped_tests": {"SomeTestPlaybook": "Some Issue"}},
+     True)
+]
+
+
+@pytest.mark.parametrize("test_playbooks, conf_dict, answer", SCRIPT_IS_SKIPPED_TEST_INPUTS)
+def test_script_has_unskipped_test_playbook(mocker, test_playbooks, conf_dict, answer):
+    """
+    Given:
+        - A Script.
+        - conf file with skipped tests.
+
+    When: running is_valid_file_in_conf_json specifically on the script.
+
+    Then: Validate the script has at least one unskipped test playbook.
+    """
+    mocker.patch.object(ConfJsonValidator, 'load_conf_file', return_value=conf_dict)
+
+    validator = ConfJsonValidator()
+
+    current = {"commonfields": {"id": "SomeScript"}, "tests": test_playbooks}
+    assert validator.is_valid_file_in_conf_json(current_file=current,
+                                                file_type=FileType.SCRIPT,
+                                                file_path="SomeFilePath") is answer
+
+
+def test_non_testable_entity_is_vaild_in_conf(mocker):
+    """
+    Given:
+        - A content entity that cant have test playbooks, specifically a test playbook
+        - Some conf file.
+
+    When: running is_valid_file_in_conf_json specifically on the content entity.
+
+    Then: Validate the content entity is valid in conf file.
+    """
+    mocker.patch.object(ConfJsonValidator, 'load_conf_file',
+                        return_value={"skipped_tests": {"SomeTestPlaybook": "Some Issue"}})
+
+    validator = ConfJsonValidator()
+
+    current = {"id": "TheTestPlaybook"}
+    assert validator.is_valid_file_in_conf_json(current_file=current,
+                                                file_type=FileType.TEST_PLAYBOOK,
+                                                file_path="SomeFilePath")